+++ R/GetColors.R
@@ -12,15 +12,15 @@
 #' @param alpha 'numeric' number.
 #'   Alpha transparency, values range from 0 (fully transparent) to 1 (fully opaque).
 #'   Specify as \code{NULL} to exclude the alpha channel value from colors.
-#' @param start,end 'numeric' number.
-#'   Starting and ending color level in the palette, respectively.
-#'   Specified as a number in the interval from 0 to 1.
-#'   Applies only to interpolated color schemes.
+#' @param stops 'numeric' vector of length 2.
+#'   Color stops defined by interval endpoints (between 0 and 1)
+#'   and used to select a subset of the color palette.
+#'   Only suitable for schemes that allow for color interpolations.
 #' @param bias 'numeric' number.
 #'   Interpolation bias where larger values result in more widely spaced colors at the high end.
 #'   See \code{\link[grDevices]{colorRamp}} function for details.
 #' @param reverse 'logical' flag.
-#'   Whether to reverse the color order in the palette.
+#'   Whether to reverse the order of colors in the scheme.
 #' @param blind 'character' string.
 #'   Type of color blindness to simulate: specify \code{"deutan"} for green-blind vision,
 #'   \code{"protan"} for red-blind vision, \code{"tritan"} for green-blue-blind vision, or
@@ -31,14 +31,20 @@
 #' @param gray 'logical' flag.
 #'   Whether to subset/reorder the \code{"bright"}, \code{"high-contrast"}, \code{"vibrant"},
 #'   and \code{"muted"} schemes to work well after conversion to gray scale.
+#' @param ...
+#'   Not used
 #'
-#' @details The suggested data type for color schemes and the characteristics of generated palettes are given in the table below.
-#'   [\bold{Type}: is the type of data being represented, either qualitative, diverging, or sequential.
+#' @details The suggested data type for color schemes and the
+#'   characteristics of generated palettes are given in the table below.
+#'   [\bold{Type}: is the type of data being represented,
+#'   either qualitative, diverging, or sequential.
 #'   \bold{Max n}: is the maximum number of colors in a generated palette.
-#'   And the maximum \code{n} value when palette colors are designed for gray-scale conversion is enclosed in parentheses.
-#'   \bold{N}: not-a-number color.
-#'   \bold{B}: background color.
-#'   \bold{F}: foreground color.
+#'   And the maximum \code{n} value when scheme colors are designed for
+#'   gray-scale conversion is enclosed in parentheses.
+#'   A value of infinity indicates that the scheme allows for color interpolations.
+#'   \bold{N}: is the not-a-number color.
+#'   \bold{B}: is the background color.
+#'   \bold{F}: is the foreground color.
 #'   \bold{Abbreviations}: --, not available]
 #'
 #'   \if{html}{\figure{table.svg}{alt="Table: color schemes"}}
@@ -48,23 +54,24 @@
 #'   intended to be accessed in their entirety and subset using vector element names.
 #'
 #' @return When argument \code{n} is specified the function
-#'   returns an object of class 'inlcol' that inherits behavior from the 'character' class.
+#'   returns an object of class 'inlpal' that inherits behavior from the 'character' class.
 #'   And when \code{n} is unspecified a variant of the \code{GetColors} function is
-#'   returned that has default (formal) argument values set equal to the values specified by the user.
+#'   returned that has default argument values set equal to the values specified by the user.
 #'
-#'   The inlcol-class object is comprised of a 'character' vector of \code{n} colors in the RGB color system.
+#'   The inlpal-class object is comprised of a 'character'
+#'   vector of \code{n} colors in the RGB color system.
 #'   Colors are specified with a string of the form \code{"#RRGGBB"} or \code{"#RRGGBBAA"}
 #'   where \code{RR}, \code{GG}, \code{BB}, and \code{AA} are the
 #'   red, green, blue, and alpha hexadecimal values (00 to FF), respectively.
 #'   Attributes of the returned object include:
 #'   \code{"names"}, the informal names assigned to colors in the palette,
 #'   where \code{NULL} indicates no color names are specified;
-#'   \code{"NaN"}, a 'character' string giving the color meant for missing data, in hexadecimal format,
-#'   where \code{NA} indicates no color is specified; and
-#'   \code{"call"}, an object of class '\link{call}' giving the unevaluated function call (expression)
-#'   that can be used to reproduce the color palette.
+#'   \code{"NaN"}, a 'character' string giving the color meant for missing data,
+#'   in hexadecimal format, where \code{NA} indicates no color is specified; and
+#'   \code{"call"}, an object of class '\link{call}' giving the unevaluated function
+#'   call (expression) that can be used to reproduce the color palette.
 #'   Use the \code{\link{eval}} function to evaluate the \code{"call"} argument.
-#'   A simple \code{plot} method is provided for the 'inlcol' class that
+#'   A simple \code{plot} method is provided for the 'inlpal' class that
 #'   shows a palette of colors using a sequence of shaded rectangles,
 #'   see \sQuote{Examples} section for usage.
 #'
@@ -85,7 +92,7 @@
 #'   Generic Mapping Tools: Improved version released, AGU, v. 94, no. 45, p. 409--410
 #'   doi:\href{https://doi.org/10.1002/2013EO450001}{10.1002/2013EO450001}
 #'
-#' @seealso \code{\link[grDevices]{col2rgb}}
+#' @seealso \code{\link{SetHinge}}, \code{\link[grDevices]{col2rgb}}
 #'
 #' @keywords color
 #'
@@ -96,8 +103,8 @@
 #' print(pal)
 #' plot(pal)
 #'
-#' Fun <- GetColors(scheme = "DEM screen", alpha = 0.9)
-#' filled.contour(datasets::volcano, color.palette = Fun)
+#' Pal <- GetColors(scheme = "DEM screen", alpha = 0.9)
+#' filled.contour(datasets::volcano, color.palette = Pal)
 #'
 #' # Diverging color schemes (scheme)
 #' op <- par(mfrow = c(6, 1), oma = c(0, 0, 0, 0))
@@ -140,12 +147,12 @@
 #' plot(GetColors(34, alpha = 0.2))
 #' par(op)
 #'
-#' # Color levels (start, end)
+#' # Color stops (stops)
 #' op <- par(mfrow = c(4, 1), oma = c(0, 0, 0, 0))
-#' plot(GetColors(255, start = 0.0, end = 1.0))
-#' plot(GetColors(255, start = 0.0, end = 0.5))
-#' plot(GetColors(255, start = 0.5, end = 1.0))
-#' plot(GetColors(255, start = 0.3, end = 0.9))
+#' plot(GetColors(255, stops = c(0.0, 1.0)))
+#' plot(GetColors(255, stops = c(0.0, 0.5)))
+#' plot(GetColors(255, stops = c(0.5, 1.0)))
+#' plot(GetColors(255, stops = c(0.3, 0.9)))
 #' par(op)
 #'
 #' # Interpolation bias (bias)
@@ -160,11 +167,9 @@
 #' par(op)
 #'
 #' # Reverse colors (reverse)
-#' op <- par(mfrow = c(4, 1), oma = c(0, 0, 0, 0), cex = 0.7)
+#' op <- par(mfrow = c(2, 1), oma = c(0, 0, 0, 0), cex = 0.7)
 #' plot(GetColors(10, reverse = FALSE))
 #' plot(GetColors(10, reverse = TRUE))
-#' plot(GetColors(10, reverse = FALSE, start = 0.5))
-#' plot(GetColors(10, reverse = TRUE,  start = 0.5))
 #' par(op)
 #'
 #' # Color blindness (blind)
@@ -189,8 +194,8 @@
 #' par(op)
 #'
 
-GetColors <- function(n, scheme="smooth rainbow", alpha=NULL, start=0, end=1,
-                      bias=1, reverse=FALSE, blind=NULL, gray=FALSE) {
+GetColors <- function(n, scheme="smooth rainbow", alpha=NULL, stops=c(0, 1),
+                      bias=1, reverse=FALSE, blind=NULL, gray=FALSE, ...) {
 
   if (!missing(n)) {
     checkmate::assertCount(n)
@@ -209,9 +214,13 @@ GetColors <- function(n, scheme="smooth rainbow", alpha=NULL, start=0, end=1,
     stop("gray component not available for '", scheme, "' scheme.")
 
   checkmate::assertNumber(alpha, lower=0, upper=1, finite=TRUE, null.ok=TRUE)
-  checkmate::assertNumber(start, lower=0, upper=1, finite=TRUE)
-  checkmate::assertNumber(end,   lower=0, upper=1, finite=TRUE)
-  stopifnot(start < end)
+
+  # backward compatibility
+  if (methods::hasArg("start")) stops[1] <- list(...)$start
+  if (methods::hasArg("end"))   stops[2] <- list(...)$end
+
+  checkmate::assertNumeric(stops, lower=0, upper=1, finite=TRUE, any.missing=FALSE,
+                           len=2, unique=TRUE, sorted=TRUE)
   checkmate::qassert(bias, "N1(0,)")
   checkmate::assertFlag(reverse)
   checkmate::assertString(blind, min.chars=1, null.ok=TRUE)
@@ -223,16 +232,15 @@ GetColors <- function(n, scheme="smooth rainbow", alpha=NULL, start=0, end=1,
       stop("simulating partial color blindness requires the dichromat package")
   }
 
-  if (nmax < Inf && (start > 0 | end < 1))
-    warning("'start' and 'end' apply only to interpolated color schemes")
+  if (nmax < Inf && !identical(stops, c(0, 1)))
+    warning("'stops' only applies to interpolated color schemes")
 
   if (missing(n)) {
     Pal <- GetColors
     formals(Pal) <- eval(substitute(
-      alist("n"=, "scheme"=a1, "alpha"=a2, "start"=a3, "end"=a4,
-            "bias"=a5, "reverse"=a6, "blind"=a7, "gray"=a8),
-      list(a1=scheme, a2=alpha, a3=start, a4=end,
-           a5=bias, a6=reverse, a7=blind, a8=gray)
+      alist("n"=, "scheme"=a1, "alpha"=a2, "stops"=a3, "bias"=a4, "reverse"=a5,
+            "blind"=a6, "gray"=a7),
+      list(a1=scheme, a2=alpha, a3=stops, a4=bias, a5=reverse, a6=blind, a7=gray)
     ))
     return(Pal)
   }
@@ -270,14 +278,14 @@ GetColors <- function(n, scheme="smooth rainbow", alpha=NULL, start=0, end=1,
     if (reverse) color <- rev(color)
     pal <- color[1:n]
   } else {
-    value <- s$data$value
-    if (is.null(value)) value <- seq_along(s$data$color)
+    value <- if (is.null(s$data$value)) seq_along(s$data$color) else s$data$value
     value <- scales::rescale(value)
     if (reverse) {
       color <- rev(color)
       value <- rev(1 - value)
     }
-    color <- scales::gradient_n_pal(color, values=value)(seq(start, end, length.out=255))
+    x <- seq.int(stops[1], stops[2], length.out=255)
+    color <- scales::gradient_n_pal(color, values=value)(x)
     pal <- grDevices::colorRampPalette(color, bias=bias, space="Lab")(n)
   }
 
@@ -302,52 +310,40 @@ GetColors <- function(n, scheme="smooth rainbow", alpha=NULL, start=0, end=1,
   }
 
   cl <- as.call(list(quote(GetColors), "n"=n, "scheme"=scheme, "alpha"=alpha,
-                     "start"=start, "end"=end, "bias"=bias, "reverse"=reverse,
+                     "stops"=stops, "bias"=bias, "reverse"=reverse,
                      "blind"=blind, "gray"=gray))
 
-  .MakeInlcolClass(pal, nan, cl)
+  .MakeInlpalClass(pal, nan, cl)
 }
 
 #' @export
 
-# Plot function for 'inlcol' color palette
+# Plot function for 'inlpal' color palette
 
-plot.inlcol <- function(x, ..., label=TRUE) {
+plot.inlpal <- function(x, ..., label=TRUE) {
   checkmate::assertCharacter(x, any.missing=FALSE, min.len=1)
   stopifnot(all(.IsColor(x)))
   checkmate::assertFlag(label)
 
   n <- length(x)
 
-  if (label && inherits(x, "inlcol")) {
+  if (label && inherits(x, "inlpal")) {
     arg <- as.list(attr(x, "call"))
     txt <- c(paste0("n = ", n),
              paste0("scheme = '", arg$scheme, "'"),
              paste0("alpha = ", arg$alpha),
-             paste0("start = ", arg$start, ", end = ", arg$end),
+             paste0("stops = c(", arg$stops[1], ", ", arg$stops[2], ")"),
              paste0("bias = ", arg$bias),
              paste0("reverse = ", arg$reverse),
              paste0("blind = '", arg$blind, "'"),
              paste0("gray = ", arg$gray))
-    is <- c(TRUE, TRUE, !is.null(arg$alpha), arg$start > 0 | arg$end < 1,
+    is <- c(TRUE, TRUE, !is.null(arg$alpha), !identical(arg$stops, c(0, 1)),
             arg$bias != 1, arg$reverse, !is.null(arg$blind), arg$gray)
     main <- paste(txt[is], collapse=", ")
     reverse <- arg$reverse
   } else {
     main <- NULL
     reverse <- FALSE
-  }
-
-  if (label && n < 35) {  # cutoff criterion for drawing tick labels
-    border <- "#D3D3D3"
-    labels <- gsub(" ", "\n", names(x))
-    if (length(labels) == 0) {
-      labels <- seq_along(x)
-      if (reverse) labels <- rev(labels)
-    }
-  } else {
-    border <- NA
-    labels <- FALSE
   }
 
   # code adapted from example in
@@ -366,9 +362,19 @@ plot.inlcol <- function(x, ..., label=TRUE) {
   } else if (n > 1) {
     xr <- c(utils::head(xr, -1) + 1 / (2 * n), utils::tail(xr, 1))
   }
-  graphics::rect(xl, 0, xr, 1, col=x, border=border, lwd=0.25)
-  graphics::axis(1, at=0:(n - 1) / n + 1 / (2 * n), labels=labels, tick=FALSE,
-                 line=-0.5, padj=1, mgp=c(3, 0, 0), col.lab="#333333")
+  graphics::rect(xl, 0, xr, 1, col=x, border=NA)
+  if (label && n < 35) {
+    at <- 0:(n - 1) / n + 1 / (2 * n)
+    lab <- gsub(" ", "\n", names(x))
+    if (length(lab) == 0) {
+      lab <- seq_along(x)
+      if (reverse) lab <- rev(lab)
+    }
+    graphics::axis(1, at=at, labels=lab, tick=FALSE, line=-0.5, padj=1,
+                   mgp=c(3, 0, 0), col.lab="#333333")
+    v <- (0:(n - 1) / n)[-1]
+    graphics::abline(v=v, col="#D3D3D3", lwd=0.25)
+  }
   graphics::box(lwd=0.25, col="#D3D3D3")
 
   invisible()
@@ -390,16 +396,16 @@ plot.inlcol <- function(x, ..., label=TRUE) {
 }
 
 
-# Constructor function for 'inlcol' class
+# Constructor function for 'inlpal' class
 
-.MakeInlcolClass <- function(x, nan, call) {
+.MakeInlpalClass <- function(x, nan, call) {
   pattern <- "^#(\\d|[a-f]){6,8}$"
   checkmate::assertCharacter(x, pattern=pattern, ignore.case=TRUE,
                              any.missing=FALSE, min.len=1)
   checkmate::assertString(nan, na.ok=TRUE, pattern=pattern, ignore.case=TRUE)
   stopifnot(is.call(call))
-  stopifnot(all(names(formals(GetColors)) %in% names(as.list(call))))
-  structure(x, nan=nan, call=call, class=append("inlcol", class(x)))
+  stopifnot(all(names(formals(GetColors)) %in% c(names(as.list(call)), "...")))
+  structure(x, nan=nan, call=call, class=append("inlpal", class(x)))
 }