+++ R/PlotMap.R
@@ -163,7 +163,7 @@
 #' model <- gstat::gstat(id = "zinc", formula = zinc~1, locations = ~x+y, data = meuse)
 #' r <- raster::interpolate(meuse.grid, model)
 #' r <- raster::mask(r, meuse.grid)
-#' Pal <- function(n) GetColors(n, start=0.3, end=0.9)
+#' Pal <- function(n) GetColors(n, stops=c(0.3, 0.9))
 #' breaks <- seq(0, 2000, by = 200)
 #' credit <- paste("Data collected in a flood plain of the river Meuse,",
 #'                 "near the village of Stein (Netherlands),",
@@ -392,7 +392,7 @@ PlotMap <- function(r, layer=1, att=NULL, n=NULL, breaks=NULL,
     } else if (raster::is.factor(r)) {
       cols <- GetColors(n, scheme="bright")
     } else {
-      cols <- GetColors(n, start=0.3, end=0.9)
+      cols <- GetColors(n, stops=c(0.3, 0.9))
     }
   }
   if (!all(.IsColor(cols))) stop("colors are not valid")